models: Draw moby models in the 3D view again
This was broken before. Also, just a cube in the pick frame for now.
