+++ src/renderer.cpp
@@ -19,6 +19,7 @@
 #include "renderer.h"
 
 #include "app.h"
+#include "imgui_includes.h" // HSV stuff.
 
 void gl_renderer::draw_spline(const std::vector<glm::vec3>& points, const glm::mat4& mvp, const glm::vec4& colour) const{
 	
@@ -113,6 +114,118 @@ void gl_renderer::draw_cube(const glm::mat4& mvp, const glm::vec4& colour) const
 	glDrawArrays(GL_TRIANGLES, 0, 108);
 
 	glDisableVertexAttribArray(0);
+}
+
+void gl_renderer::draw_moby_model(
+		moby_model& model,
+		glm::mat4 local_to_clip,
+		array_view<GLuint> textures,
+		view_mode mode) const {
+	switch(mode) {
+		case view_mode::WIREFRAME:
+			glPolygonMode(GL_FRONT_AND_BACK, GL_LINE);
+			glUseProgram(shaders.solid_colour.id());
+			break;
+		case view_mode::TEXTURED_POLYGONS:
+			glPolygonMode(GL_FRONT_AND_BACK, GL_FILL);
+			glUseProgram(shaders.textured.id());
+			break;
+	}
+	
+	moby_model_texture_data texture = {};
+	for(std::size_t i = 0; i < model.submodels.size(); i++) {
+		moby_submodel& submodel = model.submodels[i];
+		//if(!submodel.visible_in_model_viewer) {
+		//	continue;
+		//}
+		
+		//if(params.show_vertex_indices) {
+		//	auto draw_list = ImGui::GetWindowDrawList();
+		//	for(std::size_t j = 0; j < submodel.vertices.size(); j++) {
+		//		moby_model_vertex& vert = submodel.vertices[j];
+		//		glm::vec3 proj_pos = apply_local_to_screen(glm::vec4(
+		//			vert.x / (float) INT16_MAX,
+		//			vert.y / (float) INT16_MAX,
+		//			vert.z / (float) INT16_MAX, 1.f));
+		//		if(proj_pos.z > 0.f) {
+		//			draw_list->AddText(ImVec2(proj_pos.x, proj_pos.y), 0xffffffff, int_to_hex(j).c_str());
+		//		}
+		//	}
+		//}
+		
+		if(submodel.vertices.size() == 0) {
+			continue;
+		}
+		
+		if(submodel.vertex_buffer() == 0) {
+			glGenBuffers(1, &submodel.vertex_buffer());
+			glBindBuffer(GL_ARRAY_BUFFER, submodel.vertex_buffer());
+			glBufferData(GL_ARRAY_BUFFER,
+				submodel.vertices.size() * sizeof(moby_model_vertex),
+				submodel.vertices.data(), GL_STATIC_DRAW);
+		}
+		
+		if(submodel.st_buffer() == 0) {
+			glGenBuffers(1, &submodel.st_buffer());
+			glBindBuffer(GL_ARRAY_BUFFER, submodel.st_buffer());
+			glBufferData(GL_ARRAY_BUFFER,
+				submodel.st_coords.size() * sizeof(moby_model_st),
+				submodel.st_coords.data(), GL_STATIC_DRAW);
+		}
+		
+		for(moby_subsubmodel& subsubmodel : submodel.subsubmodels) {
+			if(subsubmodel.index_buffer() == 0) {
+				glGenBuffers(1, &subsubmodel.index_buffer());
+				glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, subsubmodel.index_buffer());
+				glBufferData(GL_ELEMENT_ARRAY_BUFFER,
+					subsubmodel.indices.size(),
+					subsubmodel.indices.data(), GL_STATIC_DRAW);
+			}
+			
+			if(subsubmodel.texture) {
+				texture = *subsubmodel.texture;
+			}
+			
+			switch(mode) {
+				case view_mode::WIREFRAME: {
+					glUniformMatrix4fv(shaders.solid_colour_transform, 1, GL_FALSE, &local_to_clip[0][0]);
+					
+					glm::vec4 colour = colour_coded_submodel_index(i, model.submodels.size());
+					glUniformMatrix4fv(shaders.solid_colour_transform, 1, GL_FALSE, &local_to_clip[0][0]);
+					glUniform4f(shaders.solid_colour_rgb, colour.r, colour.g, colour.b, colour.a);
+					break;
+				}
+				case view_mode::TEXTURED_POLYGONS: {
+					glUniformMatrix4fv(shaders.textured_local_to_clip, 1, GL_FALSE, &local_to_clip[0][0]);
+					glActiveTexture(GL_TEXTURE0);
+					glBindTexture(GL_TEXTURE_2D, textures.at(texture.texture_index));
+					glUniform1i(shaders.textured_sampler, 0);
+					break;
+				}
+			}
+			
+			glEnableVertexAttribArray(0);
+			glBindBuffer(GL_ARRAY_BUFFER, submodel.vertex_buffer());
+			glVertexAttribPointer(0, 3, GL_SHORT, GL_TRUE, sizeof(moby_model_vertex), (void*) offsetof(moby_model_vertex, x));
+			
+			glEnableVertexAttribArray(1);
+			glBindBuffer(GL_ARRAY_BUFFER, submodel.st_buffer());
+			glVertexAttribPointer(1, 2, GL_SHORT, GL_TRUE, sizeof(moby_model_st), (void*) offsetof(moby_model_st, s));
+			
+			glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, subsubmodel.index_buffer());
+			glDrawElements(GL_TRIANGLES, subsubmodel.indices.size(), GL_UNSIGNED_BYTE, nullptr);
+			
+			glDisableVertexAttribArray(0);
+			glDisableVertexAttribArray(1);
+		}
+	}
+}
+
+glm::vec4 gl_renderer::colour_coded_submodel_index(std::size_t index, std::size_t submodel_count) {
+	glm::vec4 colour;
+	ImGui::ColorConvertHSVtoRGB(fmod(index / (float) submodel_count, 1.f), 1.f, 1.f, colour.r, colour.g, colour.b);
+	colour.a = 1;
+	return colour;
 }
 
 void gl_renderer::reset_camera(app* a) {
